Version 0.0.0.13
Instead of auto-updating, show a notification to inform the user and give them control of the update. In the future, this should contain a changelog instead of a generic update message.
Also testing out using only the browser-based install prompt instead of our custom prompt. This will cut down a fair bit of extra code if it works correctly.